+++ src/bin/logs.rs
@@ -21,6 +21,9 @@ async fn main() -> Result<(), Error> {
             MessageType::Unknown | MessageType::Audio | MessageType::Photo => {
                 info!("{}[{}]: {}", topic, message.message_type, message.text);
             },
+            MessageType::StreamText | MessageType::StreamAudio | MessageType::StreamPhoto => {
+                info!("{}[{}][stream_id: {}][sequence: {}][is_final: {}]: {}", topic, message.message_type, message.stream_id, message.sequence, message.is_final, message.text);
+            },
             MessageType::ModuleInfo => {
                 info!("Module Info: {}\n\t{:?}", message.text, message.params);
             }