+++ src/config_message.rs
@@ -18,6 +18,9 @@ impl ConfigMessage {
             response_topics: self.response_topics.clone().unwrap_or_else(|| default.response_topics.clone()),
             sender: self.sender.clone().unwrap_or_else(|| default.sender.clone()),
             message_type: self.message_type.clone().unwrap_or_else(|| default.message_type.clone()),
+            is_final: default.is_final,
+            stream_id: default.stream_id.clone(),
+            sequence: default.sequence,
             params: default.params.clone(),
         }
     }